requirements.txt
pytest>=8.2.2
bcrypt>=4.1.3
ttkbootstrap==1.10.1
pillow>=10.0.3
reportlab>=4.2.0
loguru>=0.7.2
